pyproject.toml
[tool.poetry]
name = "car_api"
version = "0.1.0"
description = ""
authors = ["Divaltor <kuklavudu78@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
python-dotenv = "^0.19.2"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
psycopg2-binary = "^2.9.3"
django-extensions = "^3.1.5"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^6.1.0"}
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
